use lexical over dynamic scope to link expect with enclosing test
I originally avoided this approach because of the bad error messages
if you use (expect) outside of a (test) block. But I think the simplicity is
worth it. Now instead of a friendly reminder, you will get the rather ugly:

compile error: unknown symbol EXPECT_MUST_OCCUR_WITHIN_TEST_00000U

Which is not as nice, but I think the simplification is worth it. After all,
it's not a mistake people that people are likely to make in the first place.
Especially since I am the only person using this library.