use working-directory relative imports
This fixes a bug where a module could be imported multiple times, if one of
the imports used /cwd-relative paths instead of ./file-relative paths.

I could have sworn that this didn't work when I started on Judge v2, but it
appears to be work great, so I don't know what I was doing wrong before.