fix: preserve file permissions when accepting test changes
This commit fixes a bug causing `judge --accept` to reset the file
permissions of any modified test files back to the default of `0666 &
umask`.

This happened because Judge first writes to a new `foo.janet.tested`
file with default file permissions and then conditionally replaces
`foo.janet` with it. The fix is to update the permissions of the
`.tested` file to match the original one immediately before performing
the move.